formatPerformersToRanked helper function
