+++ src/components/RankingList/RankingList.stories.tsx
@@ -1,15 +1,20 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
 import RankingList from "./RankingList";
 import allPerformers from "../../mocks/allPerformers.json" with {type: "json"}
+import { formatPerformersToRanked } from "@/helpers/gameplay";
+
+const stashPerformers = allPerformers.data.findPerformers.performers.map(p => ({
+        ...p,
+        id: +p.id
+    }))
+
+const rankedPerformers = formatPerformersToRanked(stashPerformers)
 
 const meta = {
   title: "Components/RankingList",
   component: RankingList,
   args: {
-    performers: allPerformers.data.findPerformers.performers.map (p => ({
-        ...p,
-        id: +p.id
-    })),
+    performers: rankedPerformers,
     sessionHistory: [new Date("2025-06-15T16:14:53.809Z")],
   },
   parameters: {
